Create edit.src
This was a trsdos utility that I wrote while learning assembly on my Model III.  I don't remember where the password decoder code came from, I think it was a book.  The utility also used Penram, a utility that was published in the January 83 issue of 80 Micro (pg 276).  That can be found here: https://archive.org/stream/80-microcomputing-magazine-1983-01/80Microcomputing_0183#page/n275/mode/2up

There aren't any real comments for this, that was a project that was actually in progress but life happened and it never did.  I had planned on re-writing it and adding comments.  There was a secret menu that allowed for removing backup protection and changing disk serial numbers.

The code is a mess, but the fact it still exists, is a minor miracle.  The code went from my TRS-80 to my Apple //c over a serial port.  When the //c was replaced by a //gs, it went to a 3.5 inch disk where it sat for the last 30+ years.  I recently bought a Apple //gs on ebay and found the old disks and these files.  The disks were imaged to a sd card that was hooked up to the //gs thru FloppyEmu.  The data on the sd card was then copied to the PC.